fix: have fetchWithTimeout configurable
